% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a poisson glmer}
\usage{
fit_model(formula, dataset)
}
\arguments{
\item{formula}{the glmer formula}

\item{dataset}{the dataset}
}
\description{
Fit a poisson glmer
}

