% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit_model_ci.R
\name{fit_model_ci}
\alias{fit_model_ci}
\title{Fit a poisson glmer with confidence intervals on model parameters}
\usage{
fit_model_ci(formula, dataset)
}
\arguments{
\item{formula}{the glmer formula}

\item{dataset}{the dataset}
}
\description{
Fit a poisson glmer with confidence intervals on model parameters
}

